NEAT update
VAE.py : reduce memory usage for generate_render method
constants.py : new arrays, one with the name of the levels on wich the networks and neat will train on
    the other contains the end's distance of the levels
neat_sonic.py : new fitness function that takes in account the numbers of steps, some code refactoring and performances improvements
procedure.py : update for multi-level data generation

+++ models/VAE.py
@@ -212,7 +212,7 @@ class VAE():
 		:param save_path: If not None, we save the VAE's images
 		:return:
 		'''
-		images = np.load(data_path)
+		images = np.load(data_path).astype(np.float16)
 		images = images / 255
 
 		generated_images = []
